Fix incorrect use of usize for 32 bit systems
When compiling to wasm, I noticed this library would not produce the correct outputs. After a lot of frustration, I finally tracked this issue down to the line:

```
    const UTF8_WIDTH: usize = 0x4322_0000_1111_1111;
    (UTF8_WIDTH >> ((c >> 4) * 4)) & 0xf
```

This works fine on 64 bit systems, but on 32 bit systems, the `usize` type is 32 bits. If you are compiling this library directly, you do get an error about this but apparently you don't if you're using it as a dependency